Move supporting layers to separate component
