+++ app/components/map-from-id.js
@@ -2,10 +2,6 @@ import Component from '@ember/component';
 import mapboxgl from 'mapbox-gl';
 import { computed } from 'ember-decorators/object';
 import numeral from 'numeral';
-import carto from 'ember-jane-maps/utils/carto';
-
-import railConfig from '../supporting-layers/rail';
-import aerialsConfig from '../supporting-layers/aerials';
 
 export default Component.extend({
   classNameBindings: ['narrativeVisible:narrative-visible'],
@@ -19,12 +15,6 @@ export default Component.extend({
   center: [-73.869324, 40.815888],
 
   highlightedFeature: null,
-  railVisible: false,
-  railConfig,
-  aerialsConfig,
-  railSource: null,
-
-  aerialVisible: false,
 
   popup: new mapboxgl.Popup({
     closeButton: false,
@@ -156,29 +146,6 @@ export default Component.extend({
       }
 
       map.getCanvas().style.cursor = feature ? 'pointer' : '';
-    },
-
-    toggleRail() {
-      // set railSource if user is toggling rails on.
-      if (!this.get('railVisible')) {
-        const source = this.get('railConfig.source');
-        carto.getVectorTileTemplate(source['source-layers'])
-          .then(template => ({
-            id: source.id,
-            type: 'vector',
-            tiles: [template],
-          }))
-          .then((builtSource) => {
-            this.set('railSource', builtSource);
-          });
-      }
-
-
-      this.toggleProperty('railVisible');
-    },
-
-    toggleAerial() {
-      this.toggleProperty('aerialVisible');
     },
   },
 });
