+++ tests/integration/components/supporting-layers-test.js
@@ -1,0 +1,24 @@
+import { moduleForComponent, test } from 'ember-qunit';
+import hbs from 'htmlbars-inline-precompile';
+
+moduleForComponent('supporting-layers', 'Integration | Component | supporting layers', {
+  integration: true
+});
+
+test('it renders', function(assert) {
+  // Set any properties with this.set('myProperty', 'value');
+  // Handle any actions with this.on('myAction', function(val) { ... });
+
+  this.render(hbs`{{supporting-layers}}`);
+
+  assert.equal(this.$().text().trim(), '');
+
+  // Template block usage:
+  this.render(hbs`
+    {{#supporting-layers}}
+      template block text
+    {{/supporting-layers}}
+  `);
+
+  assert.equal(this.$().text().trim(), 'template block text');
+});
